Time: 0 ms (100.00%), Space: 6.4 MB (7.16%) - LeetHub

+++ 0880-decoded-string-at-index/0880-decoded-string-at-index.cpp
@@ -3,7 +3,7 @@ public:
     string decodeAtIndex(string s, int k) {
         int n=s.size();
         string res="";
-        long long len=0;
+        long len=0;
         for(auto ch:s){
             if(ch>='a' && ch<='z') len++;
             else len*= (ch-'0');
